Handle Deleting Missing Product

+++ backend/controllers/product.controller.js
@@ -70,6 +70,13 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({
+            success: false,
+            message: "Product Not Found"
+        });
+    }
     
     try {
         await Product.findByIdAndDelete(id);
@@ -79,9 +86,9 @@ export const deleteProduct = async (req, res) => {
         });
     } catch (error) {
         console.error("Error while deleting product", error.message);
-        return res.status(404).json({
+        return res.status(500).json({
             success: false,
-            message: "Product Not Found"
+            message: "Server Error"
         });
     }
 }
